requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
Bottleneck==1.3.5
braceexpand==0.1.7
cachetools==5.3.1
calmsize==0.1.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cvxpy==1.3.2
cycler==0.11.0
cymem==2.0.8
de-core-news-sm==3.7.0
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
docker-pycreds==0.4.0
ecos==2.0.12
einops==0.7.0
en-core-web-sm==3.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.10
GitPython==3.1.32
gmpy2==2.1.2
h5py==3.9.0
huggingface-hub==0.20.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.1
iopath==0.1.10
ipykernel==6.29.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jstyleson==0.0.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langcodes==3.3.0
lightning-utilities==0.9.0
llvmlite==0.41.0
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.4
numpy==1.24.4
omegaconf==2.3.0
opacus==1.1.0
opencv-python==4.9.0.80
openpyxl==3.1.2
osqp==0.6.3
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
portalocker==2.8.2
premailer==3.10.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.1
scikit-learn==1.3.0
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.29.2
setproctitle==1.3.2
setuptools==68.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
thinc==8.2.2
threadpoolctl==3.2.0
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.0
torch==1.8.0
torchaudio==0.9.0
torchdata==0.7.1
torchinfo==1.8.0
torchmetrics==1.1.2
torchtext==0.9.0
torchvision==0.10.0+cu111
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.4
vpdb==1.0.0
wandb==0.15.8
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.38.4
xgboost==2.0.2
yagmail==0.15.293
zipp==3.11.0
